@@ -33,25 +33,27 @@
   "author": "Milton Candelero <miltoncandelero@gmail.com>",
   "license": "MIT",
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^21.0.1",
-    "@rollup/plugin-node-resolve": "^13.0.4",
-    "@rollup/plugin-sucrase": "^4.0.0",
-    "@types/jest": "^27.0.2",
-    "@typescript-eslint/eslint-plugin": "^5.2.0",
-    "@typescript-eslint/parser": "^5.2.0",
-    "eslint": "^8.1.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-jest": "^25.2.2",
-    "eslint-plugin-prettier": "^4.0.0",
-    "jest": "^27.3.1",
-    "prettier": "^2.4.1",
+    "@microsoft/api-extractor": "^7.33.7",
+    "@rollup/plugin-commonjs": "^24.0.0",
+    "@rollup/plugin-node-resolve": "^15.0.1",
+    "@rollup/plugin-sucrase": "^5.0.1",
+    "@types/jest": "^29.2.4",
+    "@typescript-eslint/eslint-plugin": "^5.47.0",
+    "@typescript-eslint/parser": "^5.47.0",
+    "eslint": "^8.30.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-jest": "^27.1.7",
+    "eslint-plugin-prettier": "^4.2.1",
+    "jest": "^29.3.1",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.8.1",
     "rimraf": "^3.0.2",
-    "rollup": "^2.56.2",
+    "rollup": "^3.7.5",
     "rollup-plugin-sourcemaps": "^0.6.2",
     "rollup-plugin-terser": "^7.0.0",
-    "ts-jest": "^27.0.7",
-    "typedoc": "^0.22.7",
-    "typescript": "^4.4.4"
+    "ts-jest": "^29.0.3",
+    "typedoc": "^0.23.23",
+    "typescript": "^4.9.4"
   },
   "scripts": {
     "lint": "run-p lint:*",
@@ -68,9 +70,5 @@
     "prepublishOnly": "run-s test build doc",
     "test": "jest",
     "doc": "typedoc src/index.ts"
-  },
-  "dependencies": {
-    "@microsoft/api-extractor": "^7.18.17",
-    "npm-run-all": "^4.1.5"
   }
 }
